Move calculation files to their own folders
