+++ snap_financial_factors/calculations/net_income.py
